requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
fastapi ~= 0.115.6
uvicorn[standard] ~= 0.34.0
python-dotenv ~= 1.0.1
pydantic-settings ~= 2.7.0
httpx ~= 0.28.1